% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDEG.R
\name{DEG_analysis}
\alias{DEG_analysis}
\title{Calculates DE genes table for Wilcoxon rank-sum test, DESeq2 and edgeR}
\usage{
DEG_analysis(
  raw.exp,
  phenodata,
  treated,
  nontreated,
  class.column = "Class",
  adjust.method = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
    "none"),
  covariables = NULL,
  paired.samples.column = NULL
)
}
\arguments{
\item{raw.exp}{data.frame of raw counts}

\item{phenodata}{data.frame of phenotypic data for all samples}

\item{treated}{A string containing the name of class to be considered as treatment}

\item{nontreated}{A string containing the name of class to be considered as nontreatment}

\item{class.column}{A string contatining the name of column where treated and nontreated will be extracted}

\item{adjust.method}{Merhod of multiple comparison will be used for Wilcoxon rank-sum test pvalues.}

\item{covariables}{A vector list of variables should be used to covariates adjusment.}

\item{paired.samples.column}{A boolean variable to indicate paired-sample comparison}
}
\value{
A list of data.frame for Wilcoxon rank-sum test, DESeq2 and edgeR
}
\description{
Calculates DE genes table for Wilcoxon rank-sum test, DESeq2 and edgeR
}
