% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDEG.R
\name{get_DEG_table}
\alias{get_DEG_table}
\title{Export DEG tables for a specific DE method}
\usage{
get_DEG_table(listDEGs, method = c("Wilcox", "DESeq2", "edgeR"))
}
\arguments{
\item{listDEGs}{list of results from DEG_analysis}

\item{p_cutoff}{cut-off for p-value}

\item{log2fc_cutoff}{cut-off for log2 Fold-change values}

\item{padjusted}{Boolean value to use or not adjusted p-values for p_cutoff}
}
\value{
A list of data.frame for Wilcoxon rank-sum test, DESeq2 and edgeR
}
\description{
Export DEG tables for a specific DE method
}
