% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDEG.R
\name{log2fc_correlation_plot}
\alias{log2fc_correlation_plot}
\title{Export DEG tables for a specific DE method}
\usage{
log2fc_correlation_plot(
  listDEGs,
  method_one = c("Wilcox", "DESeq2", "edgeR"),
  method_two = c("Wilcox", "DESeq2", "edgeR")
)
}
\arguments{
\item{listDEGs}{list of results from DEG_analysis}

\item{method_one}{DE method for x-axis}

\item{method_two}{DE method for y-axis}
}
\value{
A scatter plot comparing two different DE methods
}
\description{
Export DEG tables for a specific DE method
}
