% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDEG.R
\name{overlap_DEGs}
\alias{overlap_DEGs}
\title{Calculates overlapping genes between Wilcox rank-sum test, DESeq2 and edgeR DE methods}
\usage{
overlap_DEGs(listDEGs, p_cutoff = 0.05, log2fc_cutoff = 1, padjusted = F)
}
\arguments{
\item{listDEGs}{list of results from DEG_analysis}

\item{p_cutoff}{cut-off for p-value}

\item{log2fc_cutoff}{cut-off for log2 Fold-change values}

\item{padjusted}{Boolean value to use or not adjusted p-values for p_cutoff}
}
\value{
A list of data.frame for Wilcoxon rank-sum test, DESeq2 and edgeR
}
\description{
Calculates overlapping genes between Wilcox rank-sum test, DESeq2 and edgeR DE methods
}
