% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDEG.R
\name{plot}
\alias{plot}
\title{Plot function}
\usage{
plot(listDEGs, type = c("up", "down"))
}
\arguments{
\item{listDEGs}{list of results from DEG_analysis}

\item{type}{plot up- or down-regulated genes}
}
\description{
Plot function
}
