% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDEG.R
\name{plot.triple.venn}
\alias{plot.triple.venn}
\title{Plot triple venn diagram}
\usage{
plot.triple.venn(
  a1 = c("a", "b", "c", "d", "e", "t", "g", "h"),
  a2 = c("x", "b", "c", "d", "e", "f", "g", "z"),
  a3 = c("y", "b", "c", "d", "x", "f", "g", "h"),
  labels = c("Group1", "Group2", "Group3")
)
}
\arguments{
\item{listDEGs}{list of results from DEG_analysis}

\item{p_cutoff}{cut-off for p-value}

\item{log2fc_cutoff}{cut-off for log2 Fold-change values}

\item{padjusted}{Boolean value to use or not adjusted p-values for p_cutoff}
}
\value{
A list of data.frame for Wilcoxon rank-sum test, DESeq2 and edgeR
}
\description{
Plot triple venn diagram
}
