% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDEG.R
\name{pvalue_distribution_plot}
\alias{pvalue_distribution_plot}
\title{Export DEG tables for a specific DE method}
\usage{
pvalue_distribution_plot(
  listDEGs,
  method = c("Wilcox", "DESeq2", "edgeR"),
  padjusted = F
)
}
\arguments{
\item{listDEGs}{list of results from DEG_analysis}

\item{padjusted}{Boolean value to use or not adjusted p-values for p_cutoff}

\item{p_cutoff}{cut-off for p-value}

\item{log2fc_cutoff}{cut-off for log2 Fold-change values}
}
\value{
A list of data.frame for Wilcoxon rank-sum test, DESeq2 and edgeR
}
\description{
Export DEG tables for a specific DE method
}
