% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDEG.R
\name{save_ranked_table}
\alias{save_ranked_table}
\title{Export DEG tables for a specific DE method}
\usage{
save_ranked_table(
  listDEGs,
  method = c("Wilcox", "DESeq2", "edgeR"),
  padjusted = FALSE
)
}
\arguments{
\item{listDEGs}{list of results from DEG_analysis}

\item{method}{DE method}

\item{padjusted}{Boolean value to use or not adjusted p-values for p_cutoff}
}
\value{
A ranked table for GSEA tools
}
\description{
Export DEG tables for a specific DE method
}
