% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDEG.R
\name{get_DEG_table}
\alias{get_DEG_table}
\title{Export DEG tables for a specific DE method}
\usage{
get_DEG_table(listDEGs, method = c("Wilcox", "DESeq2", "edgeR"))
}
\arguments{
\item{listDEGs}{list of results from DEG_analysis}

\item{method}{DE method}
}
\value{
A complete list of genes, log2FC, pvalue, and padj
}
\description{
Export DEG tables for a specific DE method
}
