% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDEG.R
\name{plot.triple.venn}
\alias{plot.triple.venn}
\title{Plot triple venn diagram}
\usage{
plot.triple.venn(
  a1 = c("a", "b", "c", "d", "e", "t", "g", "h"),
  a2 = c("x", "b", "c", "d", "e", "f", "g", "z"),
  a3 = c("y", "b", "c", "d", "x", "f", "g", "h"),
  labels = c("Group1", "Group2", "Group3")
)
}
\arguments{
\item{a1}{elements for the group 1}

\item{a2}{elements for the group 2}

\item{a3}{elements for the group 3}

\item{labels}{labels for each groups}
}
\value{
A list contains a Venn Diagram plot, and list of genes overlapped and exclusive for each group
}
\description{
Plot triple venn diagram
}
