% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDEG.R
\name{pvalue_distribution_plot}
\alias{pvalue_distribution_plot}
\title{Export DEG tables for a specific DE method}
\usage{
pvalue_distribution_plot(
  listDEGs,
  method = c("Wilcox", "DESeq2", "edgeR"),
  padjusted = F
)
}
\arguments{
\item{listDEGs}{list of results from DEG_analysis}

\item{method}{DE method}

\item{padjusted}{Boolean value to use or not adjusted p-values for p_cutoff}
}
\value{
A histogram plot for pvalues or padj
}
\description{
Export DEG tables for a specific DE method
}
